QSTiles: Set number of tile columns (2 of 2)
3, 4, or 5 columns. We don't want to get excessive here

PROBLEM: When enlarged first 2 tiles is enabled, the grid
view first row always shows COL_COUNT - 1 tiles, even though
the max number is two. Some crews have decided to tolerate this
but I can not. At the same time, a solid solution to effectively
measure the specs of the layout will require a total overhaul of
the DraggableGridView class. This is also not acceptable
at this time. What to do?

SOLUTION: We disable DraggableGridView from reflecting the enlarged
tile state. So if rows per tile is set to 4, DraggableGridView will
show just that, regardless of enlarged tile state. This has no
impact on the actual presentation in QSPanel. We simply don't treat
enlarged tiles as a special case. This is not the most elegant of
solutions. However, it is my assessment it is a reasonable compromise
in an effort to bring this killer enhancement to mainline.

Change-Id: Iab7e4266600a059b96d867c58f35d92005659060

Conflicts:

	res/values/cm_strings.xml
	res/xml/notification_drawer_settings.xml
	src/com/android/settings/cyanogenmod/NotificationDrawerSettings.java