Settings: add a way to not localize zone selections
In specific cases we may not want to display the localized time zone
name in the time zone picker.

For instance, when the system language is set to English (US), going
into the time zone picker, we can find "London." After we set the
system language to English (United Kingdom), London becomes localized to
"British Summer Time."

To avoid this behavior and to always display the unlocalized version
of the time zone name, we can now set localizeInPicker="false" on the
time zone entry desired to not be localized. Note that if a translation
is available for the selected time zone, it will still be applied.

In this patch we add this attribute to the Kiev time zone. In English
this patch has no effect. However if the system language is changed to
Ukranian and we select Kiev from the time zone picker:

   Without this patch: in the time zone picker it would use local time
   zone translations, e.g. "за схдноевропейським літнім часом"
   (Eastern European Summer Time).

   With this patch: it displays "Київ" in the time zone picker. But
   still displays the localized time zone in the Date & Time settings.

Ref: CYNGNOS-453

Change-Id: I48b9083ddb0a4af3fb74c2c16991c374d77db973
Signed-off-by: Roman Birg <roman@cyngn.com>